Separate early platform setup for SDL and ports - path setting - console setup - custom main Move SDL(2) platform to subfolder Fixup autotools handling for SDL
