+++ Makefile.am
@@ -32,8 +32,6 @@ libeasyrpg_player_a_SOURCES = \
 	src/audio_midi.h \
 	src/audio_resampler.cpp \
 	src/audio_resampler.h \
-	src/audio_sdl.cpp \
-	src/audio_sdl.h \
 	src/audio_secache.cpp \
 	src/audio_secache.h \
 	src/autobattle.cpp \
@@ -222,7 +220,7 @@ libeasyrpg_player_a_SOURCES = \
 	src/plane.h \
 	src/platform.cpp \
 	src/platform.h \
-	src/platform_clock.h \
+	src/platform/clock.h \
 	src/player.cpp \
 	src/player.h \
 	src/point.h \
@@ -291,10 +289,6 @@ libeasyrpg_player_a_SOURCES = \
 	src/screen.cpp \
 	src/screen.h \
 	src/shake.h \
-	src/sdl2_ui.cpp \
-	src/sdl2_ui.h \
-	src/sdl_ui.cpp \
-	src/sdl_ui.h \
 	src/shinonome_gothic.h \
 	src/shinonome_mincho.h \
 	src/span.h \
@@ -423,6 +417,22 @@ libeasyrpg_player_a_SOURCES = \
 	src/window_varlist.h \
 	src/external/rang.hpp
 
+if HAVE_SDL2
+libeasyrpg_player_a_SOURCES += \
+	src/platform/sdl/sdl2_ui.cpp \
+	src/platform/sdl/sdl2_ui.h \
+	src/platform/sdl/audio.cpp \
+	src/platform/sdl/audio.h
+endif
+
+if HAVE_SDL1
+libeasyrpg_player_a_SOURCES += \
+	src/platform/sdl/sdl_ui.cpp \
+	src/platform/sdl/sdl_ui.h \
+	src/platform/sdl/audio.cpp \
+	src/platform/sdl/audio.h
+endif
+
 if WANT_FMMIDI
 libeasyrpg_player_a_SOURCES += \
 	src/midiprogram.h \
@@ -451,44 +461,53 @@ libeasyrpg_player_a_SOURCES += \
 	src/platform/macos/midiout_device_coreaudio.h
 endif
 
-# These are mostly taken by CMake
+# These are used by CMake
 EXTRA_DIST += \
 	src/external/picojson.h \
-	src/platform/3ds/3ds_audio.cpp \
-	src/platform/3ds/3ds_audio.h \
-	src/platform/3ds/3ds_clock.cpp \
-	src/platform/3ds/3ds_clock.h \
-	src/platform/3ds/3ds_input_buttons.cpp \
-	src/platform/3ds/3ds_ui.cpp \
-	src/platform/3ds/3ds_ui.h \
+	src/platform/3ds/audio.cpp \
+	src/platform/3ds/audio.h \
+	src/platform/3ds/clock.cpp \
+	src/platform/3ds/clock.h \
+	src/platform/3ds/input_buttons.cpp \
+	src/platform/3ds/main.cpp \
+	src/platform/3ds/ui.cpp \
+	src/platform/3ds/ui.h \
+	src/platform/android/filesystem_saf.cpp \
+	src/platform/android/filesystem_saf.h \
 	src/platform/android/org_easyrpg_player_player_EasyRpgPlayerActivity.cpp \
 	src/platform/android/org_easyrpg_player_player_EasyRpgPlayerActivity.h \
-	src/platform/libretro/libretro_audio.cpp \
-	src/platform/libretro/libretro_audio.h \
-	src/platform/libretro/libretro_clock.cpp \
-	src/platform/libretro/libretro_clock.h \
-	src/platform/libretro/libretro_input_buttons.cpp \
-	src/platform/libretro/libretro_ui.cpp \
-	src/platform/libretro/libretro_ui.h \
+	src/platform/libretro/audio.cpp \
+	src/platform/libretro/audio.h \
+	src/platform/libretro/clock.cpp \
+	src/platform/libretro/clock.h \
+	src/platform/libretro/input_buttons.cpp \
+	src/platform/libretro/midiout_device.cpp \
+	src/platform/libretro/midiout_device.h \
+	src/platform/libretro/ui.cpp \
+	src/platform/libretro/ui.h \
 	src/platform/opendingux/opendingux_input_buttons.cpp \
 	src/platform/psp/psp_input_buttons.cpp \
-	src/platform/psvita/psp2_audio.cpp \
-	src/platform/psvita/psp2_audio.h \
-	src/platform/psvita/psp2_clock.cpp \
-	src/platform/psvita/psp2_clock.h \
-	src/platform/psvita/psp2_input_buttons.cpp \
-	src/platform/psvita/psp2_ui.cpp \
-	src/platform/psvita/psp2_ui.h \
-	src/platform/switch/switch_audio.cpp \
-	src/platform/switch/switch_audio.h \
-	src/platform/switch/switch_clock.cpp \
-	src/platform/switch/switch_clock.h \
-	src/platform/switch/switch_input_buttons.cpp \
-	src/platform/switch/switch_ui.cpp \
-	src/platform/switch/switch_ui.h \
-	src/platform/wii/wii_clock.cpp \
-	src/platform/wii/wii_clock.h \
-	src/platform/wii/wii_input_buttons.cpp \
+	src/platform/psvita/audio.cpp \
+	src/platform/psvita/audio.h \
+	src/platform/psvita/clock.cpp \
+	src/platform/psvita/clock.h \
+	src/platform/psvita/input_buttons.cpp \
+	src/platform/psvita/main.cpp \
+	src/platform/psvita/ui.cpp \
+	src/platform/psvita/ui.h \
+	src/platform/switch/audio.cpp \
+	src/platform/switch/audio.h \
+	src/platform/switch/clock.cpp \
+	src/platform/switch/clock.h \
+	src/platform/switch/input_buttons.cpp \
+	src/platform/switch/main.cpp \
+	src/platform/switch/ui.cpp \
+	src/platform/switch/ui.h \
+	src/platform/wii/clock.cpp \
+	src/platform/wii/clock.h \
+	src/platform/wii/input_buttons.cpp \
+	src/platform/wii/main.cpp \
+	src/platform/wii/main.h \
 	src/platform/windows/midiout_device_win32.cpp \
 	src/platform/windows/midiout_device_win32.h \
 	src/platform/windows/utils.cpp \
@@ -546,7 +565,7 @@ $(builddir)/.version-git: update_version
 	$(AM_V_at)$(top_srcdir)/builds/autoconf/git-version.sh > $@-t
 	$(AM_V_at)cmp -s $@ $@-t && rm $@-t || mv $@-t $@
 
-easyrpg_player_SOURCES = src/main.cpp
+easyrpg_player_SOURCES = src/platform/sdl/main.cpp
 easyrpg_player_CXXFLAGS = $(libeasyrpg_player_a_CXXFLAGS)
 easyrpg_player_LDADD = libeasyrpg-player.a libplayer-version.a \
 	$(LCF_LIBS) \
