+++ src/baseui.cpp
@@ -20,17 +20,17 @@
 #include "bitmap.h"
 
 #if USE_SDL==2
-#  include "sdl2_ui.h"
+#  include "platform/sdl/sdl2_ui.h"
 #elif USE_SDL==1
-#  include "sdl_ui.h"
+#  include "platform/sdl/sdl_ui.h"
 #elif USE_LIBRETRO
-#  include "platform/libretro/libretro_ui.h"
+#  include "platform/libretro/ui.h"
 #elif defined(__3DS__)
-#  include "platform/3ds/3ds_ui.h"
+#  include "platform/3ds/ui.h"
 #elif defined(__vita__)
-#  include "platform/psvita/psp2_ui.h"
+#  include "platform/psvita/ui.h"
 #elif defined(__SWITCH__)
-#  include "platform/switch/switch_ui.h"
+#  include "platform/switch/ui.h"
 #endif
 
 std::shared_ptr<BaseUi> DisplayUi;
