+++ src/filefinder_rtp.cpp
@@ -23,12 +23,12 @@
 #include <lcf/reader_util.h>
 
 #ifdef USE_LIBRETRO
-#   include "platform/libretro/libretro_ui.h"
+#  include "platform/libretro/ui.h"
 #endif
 
 #if defined(USE_SDL) && defined(__ANDROID__)
-#   include <jni.h>
-#   include <SDL_system.h>
+#  include <jni.h>
+#  include <SDL_system.h>
 #endif
 
 FileFinder_RTP::FileFinder_RTP(bool no_rtp, bool no_rtp_warnings, std::string rtp_path) {
@@ -110,6 +110,12 @@ FileFinder_RTP::FileFinder_RTP(bool no_rtp, bool no_rtp_warnings, std::string rt
 	// Fallback for unknown platforms
 	AddPath("/data/rtp/" + version_str);
 #endif
+
+#if (defined(PLAYER_NINTENDO) || defined(__vita__) || defined(__ANDROID__))
+	// skip environment paths
+	return;
+#endif
+
 	std::vector<std::string> env_paths;
 
 	// Windows paths are split by semicolon, Unix paths by colon
