+++ src/platform/3ds/audio.cpp
@@ -19,7 +19,7 @@
 
 #include <cstring>
 
-#include "3ds_audio.h"
+#include "audio.h"
 #include "filefinder.h"
 #include "game_clock.h"
 #include "output.h"
