+++ src/platform/3ds/clock.h
