+++ src/platform/libretro/audio.h
