+++ src/platform/libretro/clock.h
