+++ src/platform/libretro/midiout_device.h
