+++ src/platform/psvita/clock.cpp
@@ -15,7 +15,6 @@
  * along with EasyRPG Player. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "switch_clock.h"
+#include "clock.h"
 
-constexpr bool NxClock::is_steady;
-constexpr int64_t NxClock::ticks_per_sec;
+constexpr bool Psp2Clock::is_steady;
