+++ src/platform/psvita/clock.h
@@ -15,8 +15,8 @@
  * along with EasyRPG Player. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef EP_PLATFORM_PSP2_CLOCK_H
-#define EP_PLATFORM_PSP2_CLOCK_H
+#ifndef EP_PLATFORM_PSVITA_CLOCK_H
+#define EP_PLATFORM_PSVITA_CLOCK_H
 
 #include <psp2/kernel/threadmgr.h>
 #include <psp2/kernel/processmgr.h>
