+++ src/platform/switch/clock.h
