+++ src/platform/switch/main.cpp
@@ -15,17 +15,45 @@
  * along with EasyRPG Player. If not, see <http://www.gnu.org/licenses/>.
  */
 
+#include <switch.h>
+#include <unistd.h>
 #include "player.h"
-#include <cstdlib>
 
-// This is needed on Windows
-#ifdef USE_SDL
-#  include <SDL.h>
-#endif
+static int nxlinkSocket = -1;
 
-extern "C" int main(int argc, char* argv[]) {
+int main(int argc, char* argv[]) {
+	appletLockExit();
+
+	// Debug log over network
+	if(R_SUCCEEDED(socketInitializeDefault())) {
+		nxlinkSocket = nxlinkStdio();
+		if (nxlinkSocket < 0)
+			socketExit();
+	}
+
+	romfsInit();
+
+	char working_dir[256];
+	getcwd(working_dir, 255);
+	auto gamefs = FileFinder::Root().Create(working_dir);
+	if (gamefs)
+		FileFinder::SetGameFilesystem(gamefs);
+
+	// Run Player
 	Player::Init(argc, argv);
 	Player::Run();
 
+	romfsExit();
+
+	// Close debug log
+	if (nxlinkSocket >= 0) {
+		close(nxlinkSocket);
+		socketExit();
+		nxlinkSocket = -1;
+	}
+
+	// HOS will close us immediately afterwards, if requested by home menu.
+	// So no further cleanup possible.
+	appletUnlockExit();
 	return EXIT_SUCCESS;
 }
