+++ src/platform/wii/clock.cpp
@@ -14,8 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with EasyRPG Player. If not, see <http://www.gnu.org/licenses/>.
  */
+#include "clock.h"
 
-#include "3ds_clock.h"
-
-constexpr bool CtrClock::is_steady;
-constexpr int64_t CtrClock::ticks_per_sec;
+constexpr bool WiiClock::is_steady;
+constexpr int64_t WiiClock::ticks_per_sec;
