+++ src/platform/wii/input_buttons.cpp
